tests/wpt/web-platform-tests/svg/interact/inheritance.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="8000px">
  <title>Inheritance of pointer-events</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/interact.html#PointerEventsProperty"/>
    <h:meta name="assert" content="pointer-events inherits according to the spec."/>
    <h:meta name="assert" content="pointer-events has initial value auto."/>
  </metadata>
  <g id="container">
      <g id="target"></g>
  </g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/inheritance-testcommon.js"/>
  <script><![CDATA[

assert_inherited('pointer-events', 'auto', 'none');

  ]]></script>
</svg>
